fix: stabilize month overview gestures
