+++ app/src/main/java/com/kgs/calendar/ui/KgsCalendarApp.kt
@@ -3918,20 +3918,7 @@ private fun CalendarShell(
 
     Column(
         Modifier
-            .fillMaxSize()
-            .pointerInput(monthOverviewOpen, isMonthView) {
-                if (!monthOverviewOpen || isMonthView) return@pointerInput
-                detectDragGestures(
-                    onDrag = { change, dragAmount ->
-                        if (dragAmount.y != 0f) {
-                            change.consume()
-                            applyMonthOverviewDismissDrag(dragAmount.y)
-                        }
-                    },
-                    onDragEnd = { settleMonthOverviewDismissDrag() },
-                    onDragCancel = { settleMonthOverviewDismissDrag() },
-                )
-            },
+            .fillMaxSize(),
     ) {
         CalendarToolbar(
             state = state,
@@ -3981,7 +3968,7 @@ private fun CalendarShell(
                     alpha = monthOverviewAlpha * (1f - monthDismissProgress * 0.35f)
                 },
         ) {
-            MonthOverviewPanel(
+            MonthOverview(
                 month = overviewMonth,
                 state = state,
                 firstDayOfWeek = state.firstDayOfWeek,
@@ -4127,17 +4114,7 @@ private fun CalendarShell(
           Box(
               modifier = Modifier
                   .matchParentSize()
-                  .zIndex(200f)
-                  .pointerInput(Unit) {
-                      detectDragGestures(
-                          onDrag = { change, dragAmount ->
-                              change.consume()
-                              applyMonthOverviewDismissDrag(dragAmount.y)
-                          },
-                          onDragEnd = { settleMonthOverviewDismissDrag() },
-                          onDragCancel = { settleMonthOverviewDismissDrag() },
-                      )
-                  },
+                  .zIndex(200f),
           )
       }
       }
@@ -4975,464 +4952,6 @@ private fun MonthPillChip(pill: MonthPill, morphEnabled: Boolean = false) {
     }
 }
 
-@Composable
-private fun MonthOverviewPanel(
-    month: YearMonth,
-    state: CalendarUiState,
-    firstDayOfWeek: DayOfWeek,
-    onDaySelected: (LocalDate) -> Unit,
-    onMonthSelected: (YearMonth) -> Unit,
-    onMonthOffset: (Long) -> Unit,
-    onDismissDrag: (Float) -> Unit,
-    onDismissDragEnd: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val scope = rememberCoroutineScope()
-    val dragX = remember { Animatable(0f) }
-    var displayedMonthText by rememberSaveable { mutableStateOf(month.toString()) }
-    val displayedMonth = remember(displayedMonthText) { YearMonth.parse(displayedMonthText) }
-    var transitionTargetMonthText by remember { mutableStateOf<String?>(null) }
-    val transitionTargetMonth = remember(transitionTargetMonthText) {
-        transitionTargetMonthText?.let(YearMonth::parse)
-    }
-    var transitionDirection by remember { mutableFloatStateOf(1f) }
-    var suppressNextExternalMonthAnimation by remember { mutableStateOf(false) }
-    var monthViewportWidthPx by remember { mutableFloatStateOf(0f) }
-    var monthDragVelocityX by remember { mutableFloatStateOf(0f) }
-    var monthDragY by remember { mutableFloatStateOf(0f) }
-    var lastMonthDragNanos by remember { mutableStateOf(0L) }
-    var monthGestureMoved by remember { mutableStateOf(false) }
-    val locale = LocalAppLocale.current
-    val monthFormatter = remember(locale) { DateTimeFormatter.ofPattern("MMM", locale) }
-
-    LaunchedEffect(month) {
-        if (month.toString() == displayedMonthText) return@LaunchedEffect
-        if (suppressNextExternalMonthAnimation || monthViewportWidthPx <= 0f) {
-            suppressNextExternalMonthAnimation = false
-            displayedMonthText = month.toString()
-            dragX.snapTo(0f)
-            return@LaunchedEffect
-        }
-        val direction = if (ChronoUnit.MONTHS.between(displayedMonth, month) >= 0) 1f else -1f
-        transitionDirection = direction
-        transitionTargetMonthText = month.toString()
-        dragX.stop()
-        dragX.snapTo(0f)
-        dragX.animateTo(
-            targetValue = -direction * monthViewportWidthPx,
-            animationSpec = tween(230, easing = MotionEmphasized),
-        )
-        displayedMonthText = month.toString()
-        transitionTargetMonthText = null
-        dragX.snapTo(0f)
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .zIndex(25f)
-            .background(MaterialTheme.colorScheme.background)
-            .clickable(
-                interactionSource = remember { MutableInteractionSource() },
-                indication = null,
-                onClick = {},
-            )
-            .pointerInput(month) {
-                detectDragGestures(
-                    onDragStart = {
-                        transitionTargetMonthText = null
-                        monthDragVelocityX = 0f
-                        monthDragY = 0f
-                        monthGestureMoved = false
-                        lastMonthDragNanos = System.nanoTime()
-                        scope.launch { dragX.stop() }
-                    },
-                    onDrag = { change, dragAmount ->
-                        change.consume()
-                        if (abs(dragAmount.x) > 1.5f || abs(dragAmount.y) > 1.5f) {
-                            monthGestureMoved = true
-                        }
-                        val now = System.nanoTime()
-                        val seconds = ((now - lastMonthDragNanos).coerceAtLeast(1L) / 1_000_000_000f)
-                            .coerceAtLeast(0.001f)
-                        monthDragVelocityX = dragAmount.x / seconds
-                        monthDragY += dragAmount.y
-                        lastMonthDragNanos = now
-                        if (abs(monthDragY) > abs(dragX.value) && monthDragY < 0f) {
-                            onDismissDrag(dragAmount.y)
-                        } else {
-                            scope.launch {
-                                dragX.snapTo((dragX.value + dragAmount.x).coerceIn(-monthViewportWidthPx, monthViewportWidthPx))
-                            }
-                        }
-                    },
-                    onDragEnd = {
-                        if (abs(monthDragY) > abs(dragX.value) && monthDragY < 0f) {
-                            scope.launch { dragX.animateTo(0f, animationSpec = tween(MotionShort, easing = MotionStandard)) }
-                            onDismissDragEnd()
-                            monthGestureMoved = false
-                            return@detectDragGestures
-                        }
-                        val distanceThreshold = (monthViewportWidthPx * 0.18f).coerceAtLeast(32f)
-                        val flingThreshold = 620f
-                        val targetOffset = when {
-                            monthDragVelocityX < -flingThreshold || dragX.value < -distanceThreshold -> -monthViewportWidthPx
-                            monthDragVelocityX > flingThreshold || dragX.value > distanceThreshold -> monthViewportWidthPx
-                            else -> 0f
-                        }
-                        scope.launch {
-                            val remaining = if (monthViewportWidthPx > 0f) {
-                                (abs(targetOffset - dragX.value) / monthViewportWidthPx).coerceIn(0.15f, 1f)
-                            } else {
-                                1f
-                            }
-                            dragX.animateTo(
-                                targetOffset,
-                                animationSpec = tween(
-                                    durationMillis = (MotionMedium * remaining).roundToInt().coerceIn(MotionShort, MotionMedium),
-                                    easing = MotionEmphasized,
-                                ),
-                            )
-                            when {
-                                targetOffset < 0f -> {
-                                    suppressNextExternalMonthAnimation = true
-                                    onMonthOffset(1)
-                                }
-                                targetOffset > 0f -> {
-                                    suppressNextExternalMonthAnimation = true
-                                    onMonthOffset(-1)
-                                }
-                            }
-                            delay(80)
-                            monthGestureMoved = false
-                        }
-                    },
-                    onDragCancel = {
-                        onDismissDragEnd()
-                        monthGestureMoved = false
-                        scope.launch {
-                            dragX.animateTo(0f, animationSpec = tween(MotionShort, easing = MotionStandard))
-                        }
-                    },
-                )
-            }
-            .padding(top = 8.dp, bottom = 8.dp),
-        verticalArrangement = Arrangement.spacedBy(4.dp),
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 20.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            weekHeaderLabels(firstDayOfWeek).forEach { label ->
-                Text(
-                    text = label,
-                    modifier = Modifier.weight(1f),
-                    textAlign = TextAlign.Center,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontSize = 12.sp,
-                    lineHeight = 14.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-            }
-        }
-        BoxWithConstraints(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(month.monthGridHeight(firstDayOfWeek))
-                .clipToBounds(),
-        ) {
-            val density = LocalDensity.current
-            val widthPx = with(density) { maxWidth.toPx() }
-            LaunchedEffect(widthPx) { monthViewportWidthPx = widthPx }
-            val visibleMonths = transitionTargetMonth?.let { target ->
-                listOf(displayedMonth to 0f, target to transitionDirection)
-            } ?: listOf(
-                displayedMonth.minusMonths(1) to -1f,
-                displayedMonth to 0f,
-                displayedMonth.plusMonths(1) to 1f,
-            )
-            visibleMonths.forEach { (visibleMonth, slotOffset) ->
-                Box(
-                    modifier = Modifier
-                        .width(maxWidth)
-                        .offset { IntOffset((slotOffset * widthPx + dragX.value).roundToInt(), 0) },
-                ) {
-                    MonthGrid(
-                        month = visibleMonth,
-                        markersByDay = monthMarkersFor(
-                            visibleMonth,
-                            state.events,
-                            if (state.showCompletedTasksInCalendar) state.datedTasks else state.datedTasks.filterNot { it.isCompleted },
-                            state.taskColorMode,
-                        ),
-                        firstDayOfWeek = firstDayOfWeek,
-                        selectedStart = state.selectedDate,
-                        daysEnabled = !monthGestureMoved && abs(dragX.value) < 1f && transitionTargetMonth == null,
-                        onDaySelected = onDaySelected,
-                    )
-                }
-            }
-        }
-        Spacer(Modifier.height(8.dp))
-        MonthStrip(
-            month = month,
-            formatter = monthFormatter,
-            onMonthSelected = onMonthSelected,
-        )
-    }
-}
-
-@Composable
-private fun MonthGrid(
-    month: YearMonth,
-    markersByDay: Map<LocalDate, MonthDayMarkers>,
-    firstDayOfWeek: DayOfWeek,
-    selectedStart: LocalDate,
-    daysEnabled: Boolean = true,
-    onDaySelected: (LocalDate) -> Unit,
-) {
-    val firstDay = month.atDay(1)
-    val leadingEmptyDays = firstDay.leadingDaysFrom(firstDayOfWeek)
-    val rowCount = month.monthGridRowCount(firstDayOfWeek)
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 20.dp),
-        verticalArrangement = Arrangement.spacedBy(2.dp),
-    ) {
-        repeat(rowCount) { row ->
-            Row(Modifier.fillMaxWidth()) {
-                repeat(7) { column ->
-                    val dayIndex = row * 7 + column - leadingEmptyDays + 1
-                    if (dayIndex in 1..month.lengthOfMonth()) {
-                        val day = month.atDay(dayIndex)
-                        val markers = markersByDay[day]
-                        MonthDayCell(
-                            day = day,
-                            colors = markers?.colors.orEmpty().map(::Color),
-                            hasMore = markers?.hasMore == true,
-                            isToday = day == LocalCalendarTimeSnapshot.current.today,
-                            isSelectedStart = day == selectedStart,
-                            enabled = daysEnabled,
-                            onClick = { onDaySelected(day) },
-                            modifier = Modifier.weight(1f),
-                        )
-                    } else {
-                        Spacer(Modifier.weight(1f).height(44.dp))
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun MonthDayCell(
-    day: LocalDate,
-    colors: List<Color>,
-    hasMore: Boolean,
-    isToday: Boolean,
-    isSelectedStart: Boolean,
-    enabled: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val background = when {
-            isToday -> WarmBrown
-            isSelectedStart -> WarmPeach
-            else -> Color.Transparent
-    }
-    val textColor = if (isToday) accentContainerContentColor() else WarmInk
-    Column(
-        modifier = modifier
-            .height(44.dp)
-            .clip(RoundedCornerShape(18.dp))
-            .clickable(enabled = enabled, onClick = onClick),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Box(
-            modifier = Modifier
-                .size(34.dp)
-                .clip(CircleShape)
-                .background(background)
-                .then(
-                    if (isSelectedStart && !isToday) {
-                        Modifier.border(1.2.dp, WarmBrown.copy(alpha = 0.58f), CircleShape)
-                    } else {
-                        Modifier
-                    },
-                ),
-            contentAlignment = Alignment.Center,
-        ) {
-            Text(
-                text = day.dayOfMonth.toString(),
-                color = textColor,
-                fontSize = 14.sp,
-                lineHeight = 16.sp,
-                fontWeight = if (isToday || isSelectedStart) FontWeight.SemiBold else FontWeight.Normal,
-            )
-        }
-        Row(
-            modifier = Modifier.height(8.dp),
-            horizontalArrangement = Arrangement.spacedBy(2.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            colors.forEach { color ->
-                Box(
-                    Modifier
-                        .size(4.dp)
-                        .clip(CircleShape)
-                        .background(color),
-                )
-            }
-            if (hasMore) {
-                Text(
-                    "+",
-                    color = WarmInk,
-                    fontSize = 8.sp,
-                    lineHeight = 8.sp,
-                    modifier = Modifier.offset(y = (-2).dp),
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun MonthStrip(
-    month: YearMonth,
-    formatter: DateTimeFormatter,
-    onMonthSelected: (YearMonth) -> Unit,
-) {
-    val selectedPage = month.toMonthPage()
-    val locale = LocalAppLocale.current
-    val density = LocalDensity.current
-    val edgeScrollOffsetPx = with(density) { 20.dp.roundToPx() }
-    val listState = rememberLazyListState(
-        initialFirstVisibleItemIndex = (selectedPage - 2).coerceIn(0, MonthStripPageCount - 1),
-        initialFirstVisibleItemScrollOffset = edgeScrollOffsetPx,
-    )
-
-    LaunchedEffect(selectedPage, edgeScrollOffsetPx) {
-        val target = (selectedPage - 2).coerceIn(0, MonthStripPageCount - 1)
-        if (listState.firstVisibleItemIndex != target || abs(listState.firstVisibleItemScrollOffset - edgeScrollOffsetPx) > 2) {
-            listState.animateScrollToItem(
-                index = target,
-                scrollOffset = edgeScrollOffsetPx,
-            )
-        }
-    }
-
-    LazyRow(
-        state = listState,
-        horizontalArrangement = Arrangement.spacedBy(3.dp),
-        contentPadding = PaddingValues(horizontal = 0.dp),
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(44.dp)
-            .padding(top = 5.dp, bottom = 5.dp)
-            .horizontalEdgeFade(edgeWidth = 8.dp),
-    ) {
-        items(MonthStripPageCount, key = { it }) { page ->
-        val itemMonth = page.toMonth()
-        val selected = itemMonth == month
-        Row(
-            modifier = Modifier
-                .height(34.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(4.dp, Alignment.CenterHorizontally),
-        ) {
-            if (itemMonth.monthValue == 1) {
-                Text(
-                    itemMonth.year.toString(),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontSize = 15.sp,
-                    lineHeight = 20.sp,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1,
-                    softWrap = false,
-                    modifier = Modifier
-                        .padding(horizontal = 8.dp)
-                        .align(Alignment.CenterVertically),
-                )
-            }
-            Surface(
-                onClick = { onMonthSelected(itemMonth) },
-                modifier = Modifier.height(33.dp),
-                shape = RoundedCornerShape(16.dp),
-                color = if (selected) WarmPeach else MaterialTheme.colorScheme.surface.copy(alpha = 0.72f),
-            ) {
-                Box(
-                    modifier = Modifier.padding(horizontal = 11.dp, vertical = 3.dp),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Text(
-                        text = itemMonth.shortMonthLabel(formatter, locale),
-                        color = WarmInk,
-                        fontSize = 15.sp,
-                        lineHeight = 20.sp,
-                        fontWeight = if (selected) FontWeight.SemiBold else FontWeight.Normal,
-                        maxLines = 1,
-                        softWrap = false,
-                    )
-                }
-            }
-        }
-        }
-    }
-}
-
-@Composable
-private fun YearStrip(
-    selectedYear: Int,
-    onYearSelected: (Int) -> Unit,
-) {
-    val selectedIndex = (selectedYear - YearStripBase).coerceIn(0, YearStripPageCount - 1)
-    val listState = rememberLazyListState(initialFirstVisibleItemIndex = (selectedIndex - 2).coerceAtLeast(0))
-    LaunchedEffect(selectedIndex) {
-        listState.animateScrollToItem((selectedIndex - 2).coerceAtLeast(0))
-    }
-    LazyRow(
-        state = listState,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(52.dp)
-            .background(MaterialTheme.colorScheme.background)
-            .horizontalEdgeFade(edgeWidth = 12.dp),
-        contentPadding = PaddingValues(horizontal = 6.dp, vertical = 7.dp),
-        horizontalArrangement = Arrangement.spacedBy(5.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        items(YearStripPageCount, key = { it }) { index ->
-            val year = YearStripBase + index
-            val selected = year == selectedYear
-            Surface(
-                onClick = { onYearSelected(year) },
-                modifier = Modifier.height(36.dp),
-                shape = RoundedCornerShape(18.dp),
-                color = if (selected) WarmPeach else MaterialTheme.colorScheme.surface.copy(alpha = 0.78f),
-                border = BorderStroke(1.dp, if (selected) WarmBrown.copy(alpha = 0.42f) else WarmLine.copy(alpha = 0.62f)),
-            ) {
-                Box(
-                    modifier = Modifier.padding(horizontal = 13.dp),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Text(
-                        year.toString(),
-                        color = WarmInk,
-                        fontSize = 14.sp,
-                        lineHeight = 17.sp,
-                        fontWeight = if (selected) FontWeight.SemiBold else FontWeight.Normal,
-                    )
-                }
-            }
-        }
-    }
-}
 
 internal val LocalSheetHeaderDragModifier = compositionLocalOf<Modifier> { Modifier }
 private val LocalPendingMutations = compositionLocalOf<List<PendingMutationEntity>> { emptyList() }
@@ -11163,7 +10682,7 @@ internal fun accentContainerColor(): Color =
     }
 
 @Composable
-private fun accentContainerContentColor(): Color =
+internal fun accentContainerContentColor(): Color =
     if (MaterialTheme.colorScheme.background.isDark()) {
         MaterialTheme.colorScheme.background
     } else {
@@ -13682,7 +13201,7 @@ private fun CalendarUiState.defaultFabCreationDate(today: LocalDate = LocalDate.
         -> today
     }
 
-private fun monthMarkersFor(
+internal fun monthMarkersFor(
     month: YearMonth,
     events: List<EventEntity>,
     tasks: List<TaskEntity>,
@@ -13887,7 +13406,7 @@ internal data class DraftEventSelection(
     val allDay: Boolean = false,
 )
 
-private data class MonthDayMarkers(
+internal data class MonthDayMarkers(
     val colors: List<Int>,
     val hasMore: Boolean,
 )
